@@ -1,169 +1,172 @@
-module.exports = {
+/**
+ * Please don't forget to declare all the exports also in the fake "src/index.ts" file
+ */
+export default {
   // Components
   get ActionBar() {
-    return require('./components').ActionBar;
+    return require('./components/actionBar').default;
   },
   get ActionSheet() {
-    return require('./components').ActionSheet;
+    return require('./components/actionSheet').default;
   },
   get AnimatedImage() {
-    return require('./components').AnimatedImage;
+    return require('./components/animatedImage').default;
   },
   get AnimatedScanner() {
-    return require('./components').AnimatedScanner;
+    return require('./components/animatedScanner').default;
   },
   get Avatar() {
-    return require('./components').Avatar;
+    return require('./components/avatar').default;
   },
   get Badge() {
-    return require('./components').Badge;
+    return require('./components/badge').default;
   },
   get Button() {
-    return require('./components').Button;
+    return require('./components/button').default;
   },
   get Card() {
-    return require('./components').Card;
+    return require('./components/card').default;
   },
   get Carousel() {
-    return require('./components').Carousel;
+    return require('./components/carousel').default;
   },
   get Checkbox() {
-    return require('./components').Checkbox;
+    return require('./components/checkbox').default;
   },
   get ColorPalette() {
-    return require('./components').ColorPalette;
+    return require('./components/colorPicker/ColorPalette').default;
   },
   get ColorPicker() {
-    return require('./components').ColorPicker;
+    return require('./components/colorPicker').default;
   },
   get ColorSwatch() {
-    return require('./components').ColorSwatch;
+    return require('./components/colorPicker/ColorSwatch').default;
   },
   get ConnectionStatusBar() {
-    return require('./components').ConnectionStatusBar;
+    return require('./components/connectionStatusBar').default;
   },
   get Dialog() {
-    return require('./components').Dialog;
+    return require('./components/dialog').default;
   },
   get Drawer() {
-    return require('./components').Drawer;
+    return require('./components/drawer').default;
   },
   get FloatingButton() {
-    return require('./components').FloatingButton;
+    return require('./components/floatingButton').default;
   },
   get FeatureHighlight() {
-    return require('./components').FeatureHighlight;
+    return require('./components/featureHighlight').default;
   },
   get Hint() {
-    return require('./components').Hint;
+    return require('./components/hint').default;
   },
   get Image() {
-    return require('./components').Image;
+    return require('./components/image').default;
   },
   get KeyboardAwareScrollView() {
-    return require('./components').KeyboardAwareScrollView;
+    return require('./components/KeyboardAwareScrollView').KeyboardAwareScrollView;
   },
   get KeyboardAwareListView() {
-    return require('./components').KeyboardAwareListView;
+    return require('./components/KeyboardAwareScrollView').KeyboardAwareListView;
   },
   get BaseInput() {
-    return require('./components').BaseInput;
+    return require('./components/inputs/BaseInput').default;
   },
   get TextArea() {
-    return require('./components').TextArea;
+    return require('./components/inputs/TextArea').default;
   },
   get TextField() {
-    return require('./components').TextField;
+    return require('./components/inputs/TextField').default;
   },
   get MaskedInput() {
-    return require('./components').MaskedInput;
+    return require('./components/inputs/MaskedInput').default;
   },
   get ListItem() {
-    return require('./components').ListItem;
+    return require('./components/listItem').default;
   },
   get PageControl() {
-    return require('./components').PageControl;
+    return require('./components/pageControl').default;
   },
   get PanningProvider() {
-    return require('./components').PanningProvider;
+    return require('./components/panningViews/panningProvider').default;
   },
   get PanGestureView() {
-    return require('./components').PanGestureView;
+    return require('./components/panningViews/panGestureView').default;
   },
   get PanListenerView() {
-    return require('./components').PanListenerView;
+    return require('./components/panningViews/panListenerView').default;
   },
   get PanDismissibleView() {
-    return require('./components').PanDismissibleView;
+    return require('./components/panningViews/panDismissibleView').default;
   },
   get PanResponderView() {
-    return require('./components').PanResponderView;
+    return require('./components/panningViews/panResponderView').default;
   },
   get Picker() {
-    return require('./components').Picker;
+    return require('./components/picker').default;
   },
   get DateTimePicker() {
-    return require('./components').DateTimePicker;
+    return require('./components/dateTimePicker').default;
   },
   get ProgressBar() {
-    return require('./components').ProgressBar;
+    return require('./components/progressBar').default;
   },
   get Slider() {
-    return require('./components').Slider;
+    return require('./components/slider').default;
   },
   get GradientSlider() {
-    return require('./components').GradientSlider;
+    return require('./components/slider/GradientSlider').default;
   },
   get ColorSliderGroup() {
-    return require('./components').ColorSliderGroup;
+    return require('./components/slider/ColorSliderGroup').default;
   },
   get Stepper() {
-    return require('./components').Stepper;
+    return require('./components/stepper').default;
   },
   get TabController() {
-    return require('./components').TabController;
+    return require('./components/tabController').default;
   },
   get TabBar() {
-    return require('./components').TabBar;
+    return require('./components/tabBar').default;
   },
   get TagsInput() {
-    return require('./components').TagsInput;
+    return require('./components/tagsInput').default;
   },
   get RadioButton() {
-    return require('./components').RadioButton;
+    return require('./components/radioButton/RadioButton').default;
   },
   get RadioGroup() {
-    return require('./components').RadioGroup;
+    return require('./components/radioButton/RadioGroup').default;
   },
   get ScrollBar() {
-    return require('./components').ScrollBar;
+    return require('./components/scrollBar').default;
   },
   get SharedTransition() {
-    return require('./components').SharedTransition;
+    return require('./components/sharedTransition').default;
   },
   get StackAggregator() {
-    return require('./components').StackAggregator;
+    return require('./components/stackAggregator').default;
   },
   get Switch() {
-    return require('./components').Switch;
+    return require('./components/switch').default;
   },
   get Text() {
-    return require('./components').Text;
+    return require('./components/text').default;
   },
   get Toast() {
-    return require('./components').Toast;
+    return require('./components/toast').default;
   },
   get TouchableOpacity() {
-    return require('./components').TouchableOpacity;
+    return require('./components/touchableOpacity').default;
   },
   get View() {
-    return require('./components').View;
+    return require('./components/view').default;
   },
   get Wizard() {
-    return require('./components').Wizard;
+    return require('./components/wizard').default;
   },
   get WheelPickerDialog() {
-    return require('./components').WheelPickerDialog;
+    return require('./components/wheelPickerDialog').default;
   },
 
   // Assets
