@@ -172,16 +172,21 @@ export default class TextField extends BaseInput {
       ...this.state,
       value: props.value, // for floatingPlaceholder functionality
       floatingPlaceholderState: new Animated.Value(this.shouldFloatPlaceholder(props.value) ? 1 : 0),
-      showExpandableModal: false
+      showExpandableModal: false,
+      floatingPlaceholderTranslate: 0
     };
 
     this.generatePropsWarnings(props);
-    this.calcPlaceholderLayout();
   }
 
   UNSAFE_componentWillReceiveProps(nextProps) {
     if (nextProps.value !== this.props.value) {
-      this.setState({value: nextProps.value}, this.updateFloatingPlaceholderState);
+      this.setState({value: nextProps.value}, () => {
+        this.updateFloatingPlaceholderState();
+        if (nextProps.validateOnChange) {
+          this.validate();
+        }
+      });
     }
   }
 
@@ -191,14 +196,10 @@ export default class TextField extends BaseInput {
     }
   }
 
-  calcPlaceholderLayout = async () => {
-    if (this.shouldFakePlaceholder()) {
-      const {placeholder} = this.props;
-      const typography = this.getTypography();
-      const {width} = await Typography.measureTextSize(placeholder, typography);
-      const translate = width / 2 - (width * FLOATING_PLACEHOLDER_SCALE) / 2;
-      this.setState({floatingPlaceholderTranslate: translate / FLOATING_PLACEHOLDER_SCALE});
-    }
+  onPlaceholderLayout = (event) => {
+    const {width} = event.nativeEvent.layout;
+    const translate = width / 2 - (width * FLOATING_PLACEHOLDER_SCALE) / 2;
+    this.setState({floatingPlaceholderTranslate: translate / FLOATING_PLACEHOLDER_SCALE});
   };
 
   /** Actions */
@@ -226,9 +227,11 @@ export default class TextField extends BaseInput {
       accessibilityLabel = `${accessibilityLabel || ''}. ${value || ''}`;
     }
 
+    const accessibilityStates = this.isDisabled() ? ['disabled'] : [];
     return {
       accessibilityLabel,
-      accessibilityStates: this.isDisabled() ? ['disabled'] : []
+      // on Android accessibilityStates cause issues with expandable input
+      accessibilityStates: Constants.isIOS ? accessibilityStates : undefined
     };
   }
 
@@ -362,7 +365,7 @@ export default class TextField extends BaseInput {
     const typography = this.getTypography();
     const placeholderColor = this.getStateColor(placeholderTextColor || PLACEHOLDER_COLOR_BY_STATE.default);
 
-    if (this.shouldFakePlaceholder() && floatingPlaceholderTranslate) {
+    if (this.shouldFakePlaceholder()) {
       return (
         <View absF left>
           <Animated.Text
@@ -370,6 +373,7 @@ export default class TextField extends BaseInput {
             numberOfLines={1}
             suppressHighlighting
             accessible={false}
+            onLayout={this.onPlaceholderLayout}
             style={[
               this.styles.placeholder,
               typography,
