@@ -2,8 +2,13 @@ import _ from 'lodash';
 import PropTypes from 'prop-types';
 import React from 'react';
 import {StyleSheet, Animated, Easing} from 'react-native';
-import {Colors, Typography, PureBaseComponent, View, TouchableOpacity, Text, Image, Badge} from '../../../src';
-
+import View from '../view';
+import Text from '../text';
+import Image from '../image';
+import TouchableOpacity from '../touchableOpacity';
+import Badge from '../badge';
+import {PureBaseComponent} from '../../commons';
+import {Colors, Typography} from '../../style';
 const INDICATOR_BG_COLOR = Colors.blue30;
 const INDICATOR_HEIGHT = 2;
 const INDICATOR_SPACINGS = 16;
