@@ -8,4 +8,5 @@ export * from './interactableComponents';
 export * from './nativeComponents';
 export * from './screensComponents';
 export * from './services';
-export * from './style';
+export * from './generatedTypes';
+export * from './globalTypes';
